+++ Assets/Libraries/VoronoiLib/Structures/VPoint.cs
@@ -6,10 +6,10 @@ namespace VoronoiLib.Structures
 {
     public class VPoint
     {
-        public double X { get; }
-        public double Y { get; }
+        public float X { get; }
+        public float Y { get; }
 
-        internal VPoint(double x, double y)
+        internal VPoint(float x, float y)
         {
             X = x;
             Y = y;
